--- simulations/big_double.sumocfg
@@ -1,7 +1,0 @@
-<configuration>
-    <input>
-       <net-file value="network_sources\big_double\big_double.net.xml"/>
-       <route-files value="network_sources\big_double\big_double.rou.xml"/>
-    </input>
-	<time-to-teleport value="-1"/>
-</configuration>
