--- simulations/intersection.sumocfg
@@ -1,7 +1,0 @@
-<configuration>
-    <input>
-       <net-file value="network_sources/intersection.net.xml"/>
-       <route-files value="network_sources/intersection.rou.xml"/>
-    </input>
-	<time-to-teleport value="-1"/>
-</configuration>
