--- simulations/israel_double_intersection.sumocfg
@@ -1,7 +1,0 @@
-<configuration>
-    <input>
-       <net-file value="network_sources\networks\israel_double_intersection.net.xml"/>
-       <route-files value="network_sources\trips\israel_double_intersection.trip.xml"/>
-    </input>
-	<time-to-teleport value="-1"/>
-</configuration>
