--- simulations/network_sources/intersection.rou.xml
@@ -1,18 +1,0 @@
-<routes>
-	<vType id="car" type="passenger" length="5" accel="3.5" decel="2.2" sigma="1.0"/>
-		<flow id="carflow01" type="car" beg="0" end="9999" probability="0.10" from="-gneE0" to="gneE1" departLane="1" />
-		<flow id="carflow02" type="car" beg="0" end="9999" probability="0.06" from="-gneE0" to="gneE2" departLane="0" />
-		<flow id="carflow03" type="car" beg="0" end="9999" probability="0.06" from="-gneE0" to="gneE3" departLane="0" />
-		
-		<flow id="carflow10" type="car" beg="0" end="9999" probability="0.06" from="-gneE1" to="gneE0" departLane="0"/>
-		<flow id="carflow12" type="car" beg="0" end="9999" probability="0.10" from="-gneE1" to="gneE2" departLane="1"/>
-		<flow id="carflow13" type="car" beg="0" end="9999" probability="0.06" from="-gneE1" to="gneE3" departLane="0"/>
-		
-		<flow id="carflow21" type="car" beg="0" end="9999" probability="0.06" from="-gneE2" to="gneE1" departLane="0"/>
-		<flow id="carflow20" type="car" beg="0" end="9999" probability="0.06" from="-gneE2" to="gneE0" departLane="0"/>
-		<flow id="carflow23" type="car" beg="0" end="9999" probability="0.10" from="-gneE2" to="gneE3" departLane="1"/>
-		
-		<flow id="carflow31" type="car" beg="0" end="9999" probability="0.06" from="-gneE3" to="gneE1" departLane="0"/>
-		<flow id="carflow32" type="car" beg="0" end="9999" probability="0.06" from="-gneE3" to="gneE2" departLane="0"/>
-		<flow id="carflow30" type="car" beg="0" end="9999" probability="0.10" from="-gneE3" to="gneE0" departLane="1"/>
-</routes>
