+++ utils/replay_memory.py
@@ -23,3 +23,5 @@ class ReplayMemory(object):
 
     def __len__(self):
         return len(self.memory)
+
+
